remove hand-made wrapper script, now that setuptools does it for us
it turns out that setuptools actually installs the wrapper scripts we
were looking for, so our bup-cron wrapper isn't necessary
anymore. besides, __init__.py can now also be run as a main itself, so
we symlink to it so we can still run this from the source tree
directly.

remove other comments regarding where to find bup, installing in the
$PATH seems to be fine.